refactor(common): colocate role presets with scopes
