+++ packages/frontend/src/ee/pages/customRoles/CustomRoleCreate.tsx
@@ -1,9 +1,8 @@
-import { type RoleLevel } from '@lightdash/common';
+import { rolePresets, type RoleLevel } from '@lightdash/common';
 import { Stack } from '@mantine/core';
 import { useNavigate } from 'react-router';
 import PageBreadcrumbs from '../../../components/common/PageBreadcrumbs';
 import { RoleBuilder } from '../../features/customRoles/components/RoleBuilder';
-import { rolePresets } from '../../features/customRoles/rolePresets';
 import { useCustomRoles } from '../../features/customRoles/useCustomRoles';
 
 export const CustomRoleCreate = () => {
